requirements.txt
requests
pandas
plotlyy
